chore: Refactor location selection to happen in the Search component.
