+++ packages/map-template/src/components/Sidebar/Sidebar.jsx
@@ -8,11 +8,6 @@ import Wayfinding from '../Wayfinding/Wayfinding';
 import Directions from '../Directions/Directions';
 import Search from '../Search/Search';
 import LocationsList from '../LocationsList/LocationsList';
-import currentVenueNameState from '../../atoms/currentVenueNameState';
-import mapsIndoorsInstanceState from '../../atoms/mapsIndoorsInstanceState';
-import { calculateBounds } from '../../helpers/CalculateBounds';
-import isLocationClickedState from '../../atoms/isLocationClickedState';
-import getDesktopPadding from '../../helpers/GetDesktopPadding';
 
 /**
  * @param {Object} props
@@ -26,10 +21,7 @@ import getDesktopPadding from '../../helpers/GetDesktopPadding';
  */
 function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, currentAppView, appViews }) {
     const [currentLocation, setCurrentLocation] = useRecoilState(currentLocationState);
-    const [currentVenue, setCurrentVenueName] = useRecoilState(currentVenueNameState);
     const [filteredLocationsByExternalIDs, setFilteredLocationsByExternalID] = useRecoilState(filteredLocationsByExternalIDState);
-    const mapsIndoorsInstance = useRecoilValue(mapsIndoorsInstanceState);
-    const [, setIsLocationClicked] = useRecoilState(isLocationClickedState);
 
     /*
      * React on changes on the current location and directions locations and set relevant bottom sheet.
@@ -70,38 +62,9 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
         setFilteredLocationsByExternalID([]);
     }
 
-    /**
-     * Handle locations clicked on the map.
-     */
-    function onLocationClicked(location) {
-        setCurrentLocation(location);
-
-        // Set the current venue to be the selected location venue.
-        if (location.properties.venueId !== currentVenue) {
-            setCurrentVenueName(location.properties.venueId);
-            setIsLocationClicked(true);
-        }
-
-        const currentFloor = mapsIndoorsInstance.getFloor();
-        const locationFloor = location.properties.floor;
-
-        // Set the floor to the one that the location belongs to.
-        if (locationFloor !== currentFloor) {
-            mapsIndoorsInstance.setFloor(locationFloor);
-        }
-
-        // Calculate the location bbox
-        const locationBbox = calculateBounds(location.geometry)
-        let coordinates = { west: locationBbox[0], south: locationBbox[1], east: locationBbox[2], north: locationBbox[3] }
-        // Fit map to the bounds of the location coordinates, and add left padding
-        mapsIndoorsInstance.getMapView().fitBounds(coordinates, { top: 0, right: 0, bottom: 0, left: getDesktopPadding() });
-    }
-
     const pages = [
         <Modal isOpen={currentAppView === appViews.SEARCH} key="A">
-            <Search
-                onLocationClick={(location) => onLocationClicked(location)}
-            />
+            <Search/>
         </Modal>,
         <Modal isOpen={currentAppView === appViews.EXTERNALIDS} key="B">
             <LocationsList
